分析結果グラフの部分を別コンポーネントに分割した。
